Cleanup custom params
