+++ src/lib.rs
@@ -110,12 +110,14 @@ mod macros;
 
 mod conv;
 mod error;
-mod runtime;
 mod tag;
 mod types;
 mod value;
 
-/// Custom types
+/// Functions for initializing and interacting with the OCaml runtime
+pub mod runtime;
+
+/// Custom types, used for allocating Rust values owned by the OCaml garbage collector
 pub mod custom;
 
 pub use crate::error::{CamlError, Error};
@@ -127,42 +129,3 @@ pub use sys::mlvalues::{Intnat as Int, Uintnat as Uint};
 pub type Float = f64;
 pub use crate::custom::Custom;
 pub use crate::tag::Tag;
-
-/// Initialize the OCaml runtime, this will all command-line arguments to be available using
-/// `Sys.argv`.
-///
-/// NOTE: Your application must link the OCaml libraries to be able to use this function
-pub fn startup() {
-    let args = std::env::args()
-        .map(|arg| std::ffi::CString::new(arg).unwrap())
-        .collect::<Vec<std::ffi::CString>>();
-
-    // convert the strings to raw pointers
-    let mut c_args = args
-        .iter()
-        .map(|arg| arg.as_ptr() as *mut u8)
-        .collect::<Vec<*mut u8>>();
-    unsafe { crate::sys::callback::caml_startup(c_args.as_mut_ptr()) }
-}
-
-/// Initialize the OCaml runtime, this will all command-line arguments to be available using
-/// `Sys.argv`.
-///
-/// NOTE: Your application must link the OCaml libraries to be able to use this function
-pub fn caml_main() {
-    let args = std::env::args()
-        .map(|arg| std::ffi::CString::new(arg).unwrap())
-        .collect::<Vec<std::ffi::CString>>();
-
-    // convert the strings to raw pointers
-    let mut c_args = args
-        .iter()
-        .map(|arg| arg.as_ptr() as *mut u8)
-        .collect::<Vec<*mut u8>>();
-    unsafe { crate::sys::callback::caml_main(c_args.as_mut_ptr()) }
-}
-
-/// Shutdown and cleanup OCaml runtime
-pub fn shutdown() {
-    unsafe { crate::sys::callback::caml_shutdown() }
-}
